SenderCounted list
Implementation of a list of characters whose dialogues are counted to increase the day.
Do not increment the dialogues count when the boss is calling.
Dialogues on day 0 are moved on day 1 and their priority is raised.
Dialogue d_0008 is removed, as it is not needed anymore as third dialogue on day 0.